Update destination field to be correct
This little guy...

The big network update corrected a way in which it was doing things
wrong, which exposed this guy. It is incorrect to always ask the person
for the information of the person. That's the whole point of this
system. Previously the network, despite addressing a message to a single
person, sent it to everyone (more like received it by everyone). That's
fixed in there, so this needed to be fixed too.

On the plus side, I really trust the tests!